refactor: move files into chess dir for readability
