+++ packages/front/src/chess/chess.constants.ts
