+++ packages/front/src/chess/chess.models.ts
