+++ packages/front/src/components/ChessBoard/ChessBoard.lib.ts
@@ -1,20 +1,29 @@
 import { splitEvery } from 'ramda'
-import { BLACK, CHESS_BOARD_SIZE, DARK, LIGHT } from '../../constants'
-import { TChessBoard, TChessPieceColor, TChessSquareColor } from '../../models'
+import {
+  BLACK,
+  CHESS_BOARD_SIZE,
+  DARK,
+  LIGHT
+} from '../../chess/chess.constants'
+import {
+  TChessBoard,
+  TChessPieceColor,
+  TChessSquareColor
+} from '../../chess/chess.models'
 
 export function coordinateToSquareColor(
-	x: number,
-	y: number,
+  x: number,
+  y: number
 ): TChessSquareColor {
-	return (x + y) % 2 === 0 ? LIGHT : DARK
+  return (x + y) % 2 === 0 ? LIGHT : DARK
 }
 
 export function parseRanks(board: TChessBoard, sidePlaying: TChessPieceColor) {
-	const ranks = splitEvery(CHESS_BOARD_SIZE, board)
+  const ranks = splitEvery(CHESS_BOARD_SIZE, board)
 
-	if (sidePlaying === BLACK) {
-		return ranks.map((rank) => [...rank].reverse())
-	}
+  if (sidePlaying === BLACK) {
+    return ranks.map((rank) => [...rank].reverse())
+  }
 
-	return [...ranks].reverse()
+  return [...ranks].reverse()
 }
