+++ packages/front/src/components/ChessBoard/components/ChessBoardSquare/components/ChessPiece/ChessPiece.lib.ts
@@ -11,23 +11,23 @@ import whiteRook from '../../../../../../assets/pieces/SHAKKI_VALKONE_TORNI.svg'
 import whiteBishop from '../../../../../../assets/pieces/SHAKKI_VALKONE_LAHETTI.svg'
 import whitePawn from '../../../../../../assets/pieces/SHAKKI_VALKONE_SOTILAS.svg'
 
-import { EChessPiece } from '../../../../../../models'
+import { EChessPiece } from '../../../../../../chess/chess.models'
 
 const PIECE_TO_IMAGE_MAP: Record<EChessPiece, string> = {
-	[EChessPiece.WhitePawn]: whitePawn,
-	[EChessPiece.WhiteKnight]: whiteKnight,
-	[EChessPiece.WhiteBishop]: whiteBishop,
-	[EChessPiece.WhiteRook]: whiteRook,
-	[EChessPiece.WhiteQueen]: whiteQueen,
-	[EChessPiece.WhiteKing]: whiteKing,
-	[EChessPiece.BlackPawn]: blackPawn,
-	[EChessPiece.BlackKnight]: blackKnight,
-	[EChessPiece.BlackBishop]: blackBishop,
-	[EChessPiece.BlackRook]: blackRook,
-	[EChessPiece.BlackQueen]: blackQueen,
-	[EChessPiece.BlackKing]: blackKing,
+  [EChessPiece.WhitePawn]: whitePawn,
+  [EChessPiece.WhiteKnight]: whiteKnight,
+  [EChessPiece.WhiteBishop]: whiteBishop,
+  [EChessPiece.WhiteRook]: whiteRook,
+  [EChessPiece.WhiteQueen]: whiteQueen,
+  [EChessPiece.WhiteKing]: whiteKing,
+  [EChessPiece.BlackPawn]: blackPawn,
+  [EChessPiece.BlackKnight]: blackKnight,
+  [EChessPiece.BlackBishop]: blackBishop,
+  [EChessPiece.BlackRook]: blackRook,
+  [EChessPiece.BlackQueen]: blackQueen,
+  [EChessPiece.BlackKing]: blackKing
 }
 
 export function chessPieceToImageSrc(piece: EChessPiece): string {
-	return PIECE_TO_IMAGE_MAP[piece]
+  return PIECE_TO_IMAGE_MAP[piece]
 }
